backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openai
feedparser
python-dotenv
httpx
